backend/requirements.txt
Flask
torch
Pillow
ultralytics
transformers
langchain
langchain-community
sentence-transformers
faiss-cpu
deep-translator
